requirement.txt
wheel
psutil
fastapi
uvicorn[standard]
prometheus-client
python-daemon
gunicorn